Macros set in the file now visible after preprocessing complete. Search ranges made project-wide. Found Address list no longer resizes address section after each search.

+++ Tools/LSWindows/Src/LSWWin.h
@@ -32,7 +32,7 @@ typedef double DOUBLE;
 #endif	// #ifdef WIN32_LEAN_AND_MEAN
 
 #ifndef MAXUINT64
-#define MAXUINT64 (static_cast<uint64_t>(~0))
+#define MAXUINT64 static_cast<uint64_t>(~static_cast<uint64_t>(0))
 #endif	// #ifndef MAXUINT64
 
 #include <CommCtrl.h>
